pyproject.toml
[tool.poetry]
name = "stadsspel"
version = "0.1.0"
description = "Mobile-first city game PWA with realtime map, offline support, and admin approval workflow"
authors = ["Stadsspel Team"]
readme = "README.md"
packages = [{include = "app", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
geoalchemy2 = "^0.14.3"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
shapely = "^2.0.2"
geojson = "^3.1.0"
pillow = "^10.2.0"
aiofiles = "^23.2.1"
python-dotenv = "^1.0.0"
bcrypt = ">=4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
black = "^24.1.1"
ruff = "^0.1.14"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
migrate = "alembic upgrade head"
seed = "app.seeds.seed_demo_city:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
